增加几个接口，增加修改，查看趋势，饼图，折线图

+++ src/main/java/com/layor/tinyflow/entity/UrlListResponseDTO.java
@@ -12,6 +12,6 @@ public class UrlListResponseDTO {
     private String shortCode;     // 短链码
     private String longUrl;       // 原始长链接
     private Long totalVisits;     // 总访问次数
-    private Long todayVisits;     // 今日访问次数
+    private Integer todayVisits;     // 今日访问次数
     private LocalDateTime createdAt; // 创建时间
 }
